requirements.txt
matplotlib==3.10.8
numpy==2.0.0
torch==2.9.0
torchprofile==0.0.4
torchvision==0.24.0
wandb==0.24.2
hydra-core==1.3.2
omegaconf==2.3.0